requirements.txt
Django==4.2.7
Django REST framework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.0
redis==4.5.5
Pillow==10.1.0
celery==5.3.4
django-celery-beat==2.5.0
minio==7.1.13
python-dotenv==1.0.0